Add modules to improve structure
 - add `label_formats` module
 - add `labeling_strategies` module
 - add script for each class
 - use classmethods
