+++ labelCloud/control/drawing_manager.py
@@ -1,7 +1,8 @@
 from typing import TYPE_CHECKING, Union
 
+from labeling_strategies import BaseLabelingStrategy
+
 from .bbox_controller import BoundingBoxController
-from .drawing_strategies import IDrawingStrategy
 
 if TYPE_CHECKING:
     from view.gui import GUI
@@ -11,20 +12,20 @@ class DrawingManager(object):
     def __init__(self, bbox_controller: BoundingBoxController) -> None:
         self.bbox_controller = bbox_controller
         self.view: Union["GUI", None] = None
-        self.drawing_strategy: Union[IDrawingStrategy, None] = None
+        self.drawing_strategy: Union[BaseLabelingStrategy, None] = None
 
     def set_view(self, view: "GUI") -> None:
         self.view = view
         self.view.glWidget.drawing_mode = self
 
     def is_active(self) -> bool:
-        return isinstance(self.drawing_strategy, IDrawingStrategy)
+        return isinstance(self.drawing_strategy, BaseLabelingStrategy)
 
     def has_preview(self) -> bool:
         if self.is_active():
             return self.drawing_strategy.__class__.PREVIEW
 
-    def set_drawing_strategy(self, strategy: IDrawingStrategy) -> None:
+    def set_drawing_strategy(self, strategy: BaseLabelingStrategy) -> None:
         if self.is_active() and self.drawing_strategy == strategy:
             self.reset()
             print("Deactivated drawing!")
