+++ labelCloud/labeling_strategies/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseLabelingStrategy
+from .picking import PickingStrategy
+from .spanning import SpanningStrategy
