+++ labelCloud/model/__init__.py
@@ -1,0 +1,2 @@
+from .bbox import BBox
+from .point_cloud import PointCloud
